Moving `SearchBarForm` from `EditWidgetFrame` to `WidgetQueryControls`.
