+++ graylog2-web-interface/src/views/components/WidgetQueryControls.tsx
@@ -16,9 +16,16 @@
  */
 import * as React from 'react';
 import styled from 'styled-components';
-import { Field, useFormikContext } from 'formik';
+import { Field } from 'formik';
+import moment from 'moment';
+import { useContext } from 'react';
 
-import connect from 'stores/connect';
+import connect, { useStore } from 'stores/connect';
+import { createElasticsearchQueryString } from 'views/logic/queries/Query';
+import Widget from 'views/logic/widgets/Widget';
+import { WidgetActions } from 'views/stores/WidgetStore';
+import { DEFAULT_TIMERANGE } from 'views/Constants';
+import { SearchConfigStore } from 'views/stores/SearchConfigStore';
 import { Col, Row } from 'components/graylog';
 import { Icon } from 'components/common';
 import DocumentationLink from 'components/support/DocumentationLink';
@@ -29,12 +36,13 @@ import { StreamsStore } from 'views/stores/StreamsStore';
 import { GlobalOverrideActions, GlobalOverrideStore } from 'views/stores/GlobalOverrideStore';
 import GlobalOverride from 'views/logic/search/GlobalOverride';
 import SearchActions from 'views/actions/SearchActions';
-import type { SearchBarFormValues } from 'views/Constants';
+import WidgetContext from 'views/components/contexts/WidgetContext';
 
 import TimeRangeInput from './searchbar/TimeRangeInput';
 import StreamsFilter from './searchbar/StreamsFilter';
 import SearchButton from './searchbar/SearchButton';
 import QueryInput from './searchbar/AsyncQueryInput';
+import SearchBarForm from './searchbar/SearchBarForm';
 
 type Props = {
   availableStreams: Array<any>,
@@ -68,6 +76,17 @@ const ResetFilterButton = styled(Button)`
 
 const _resetOverride = () => GlobalOverrideActions.reset().then(SearchActions.refresh);
 
+const _onSubmit = (values, widget: Widget) => {
+  const { timerange, streams, queryString } = values;
+  const newWidget = widget.toBuilder()
+    .timerange(timerange)
+    .query(createElasticsearchQueryString(queryString))
+    .streams(streams)
+    .build();
+
+  return WidgetActions.update(widget.id, newWidget);
+};
+
 const ResetOverrideHint = () => (
   <CenteredBox>
     These controls are disabled, because a filter is applied to all widgets.{' '}
@@ -76,62 +95,75 @@ const ResetOverrideHint = () => (
 );
 
 const WidgetQueryControls = ({ availableStreams, globalOverride }: Props) => {
+  const widget = useContext(WidgetContext);
+  const config = useStore(SearchConfigStore, ({ searchesClusterConfig }) => searchesClusterConfig);
+  const limitDuration = moment.duration(config?.query_time_range_limit).asSeconds() ?? 0;
+  const { streams } = widget;
+  const timerange = widget.timerange ?? DEFAULT_TIMERANGE;
+  const { query_string: queryString } = widget.query ?? createElasticsearchQueryString('');
+
   const isGloballyOverridden: boolean = globalOverride !== undefined
     && globalOverride !== null
     && (globalOverride.query !== undefined || globalOverride.timerange !== undefined);
   const Wrapper = isGloballyOverridden ? BlurredWrapper : React.Fragment;
-  const { dirty, isValid, isSubmitting, handleSubmit, values, setFieldValue } = useFormikContext<SearchBarFormValues>();
 
   return (
     <>
       {isGloballyOverridden && <ResetOverrideHint />}
-      <Wrapper>
-        <TopRow>
-          <Col md={4}>
-            <TimeRangeInput disabled={isGloballyOverridden}
-                            onChange={(nextTimeRange) => setFieldValue('timerange', nextTimeRange)}
-                            value={values?.timerange}
-                            hasErrorOnMount={!isValid} />
-          </Col>
+      <SearchBarForm initialValues={{ timerange, streams, queryString }}
+                     limitDuration={limitDuration}
+                     onSubmit={(values) => _onSubmit(values, widget)}
+                     validateOnMount={false}>
+        {({ dirty, isValid, isSubmitting, handleSubmit, values, setFieldValue }) => (
+          <Wrapper>
+            <TopRow>
+              <Col md={4}>
+                <TimeRangeInput disabled={isGloballyOverridden}
+                                onChange={(nextTimeRange) => setFieldValue('timerange', nextTimeRange)}
+                                value={values?.timerange}
+                                hasErrorOnMount={!isValid} />
+              </Col>
 
-          <Col md={8}>
-            <Field name="streams">
-              {({ field: { name, value, onChange } }) => (
-                <StreamsFilter value={value}
-                               disabled={isGloballyOverridden}
-                               streams={availableStreams}
-                               onChange={(newStreams) => onChange({ target: { value: newStreams, name } })} />
-              )}
-            </Field>
-          </Col>
-        </TopRow>
+              <Col md={8}>
+                <Field name="streams">
+                  {({ field: { name, value, onChange } }) => (
+                    <StreamsFilter value={value}
+                                   disabled={isGloballyOverridden}
+                                   streams={availableStreams}
+                                   onChange={(newStreams) => onChange({ target: { value: newStreams, name } })} />
+                  )}
+                </Field>
+              </Col>
+            </TopRow>
 
-        <Row className="no-bm">
-          <Col md={12}>
-            <div className="pull-right search-help">
-              <DocumentationLink page={DocsHelper.PAGES.SEARCH_QUERY_LANGUAGE}
-                                 title="Search query syntax documentation"
-                                 text={<Icon name="lightbulb" type="regular" />} />
-            </div>
-            <SearchButton disabled={isGloballyOverridden || isSubmitting || !isValid}
-                          dirty={dirty} />
+            <Row className="no-bm">
+              <Col md={12}>
+                <div className="pull-right search-help">
+                  <DocumentationLink page={DocsHelper.PAGES.SEARCH_QUERY_LANGUAGE}
+                                     title="Search query syntax documentation"
+                                     text={<Icon name="lightbulb" type="regular" />} />
+                </div>
+                <SearchButton disabled={isGloballyOverridden || isSubmitting || !isValid}
+                              dirty={dirty} />
 
-            <Field name="queryString">
-              {({ field: { name, value, onChange } }) => (
-                <QueryInput value={value}
-                            disabled={isGloballyOverridden}
-                            placeholder={'Type your search query here and press enter. E.g.: ("not found" AND http) OR http_response_code:[400 TO 404]'}
-                            onChange={(newQuery) => {
-                              onChange({ target: { value: newQuery, name } });
+                <Field name="queryString">
+                  {({ field: { name, value, onChange } }) => (
+                    <QueryInput value={value}
+                                disabled={isGloballyOverridden}
+                                placeholder={'Type your search query here and press enter. E.g.: ("not found" AND http) OR http_response_code:[400 TO 404]'}
+                                onChange={(newQuery) => {
+                                  onChange({ target: { value: newQuery, name } });
 
-                              return Promise.resolve(newQuery);
-                            }}
-                            onExecute={handleSubmit as () => void} />
-              )}
-            </Field>
-          </Col>
-        </Row>
-      </Wrapper>
+                                  return Promise.resolve(newQuery);
+                                }}
+                                onExecute={handleSubmit as () => void} />
+                  )}
+                </Field>
+              </Col>
+            </Row>
+          </Wrapper>
+        )}
+      </SearchBarForm>
     </>
   );
 };
