+++ graylog2-web-interface/src/views/components/widgets/SaveOrCancelButtons.tsx
@@ -15,32 +15,19 @@
  * <http://www.mongodb.com/licensing/server-side-public-license>.
  */
 import * as React from 'react';
-import { useCallback } from 'react';
-import { useFormikContext } from 'formik';
 
 import { Button, ButtonToolbar } from 'components/graylog';
 
 type Props = {
   onCancel: () => void,
-  onFinish: (...args: any[]) => void,
+  onFinish: () => void,
 };
 
-const SaveOrCancelButtons = ({ onFinish, onCancel }: Props) => {
-  const { handleSubmit, dirty } = useFormikContext();
-  const _onFinish = useCallback((...args) => {
-    if (handleSubmit && dirty) {
-      handleSubmit();
-    }
-
-    return onFinish(...args);
-  }, [onFinish, handleSubmit, dirty]);
-
-  return (
-    <ButtonToolbar className="pull-right">
-      <Button onClick={_onFinish} bsStyle="primary">Save</Button>
-      <Button onClick={onCancel}>Cancel</Button>
-    </ButtonToolbar>
-  );
-};
+const SaveOrCancelButtons = ({ onFinish, onCancel }: Props) => (
+  <ButtonToolbar className="pull-right">
+    <Button onClick={onFinish} bsStyle="primary">Save</Button>
+    <Button onClick={onCancel}>Cancel</Button>
+  </ButtonToolbar>
+);
 
 export default SaveOrCancelButtons;
